🛠️ REFACTOR: Amélioration du système de routes backend
